Hangman Unit-7
Added function show_hidden_word which receives the secret word and the current user guesses, and returns a string with the current correct guesses of the game.

Added function check_win which receives the secret word with the current user guesses, and checks whether the user won by guessing the word.

Cleaned up the main function.